test: SaveFile UseCase custom value

+++ 04.multiplication/src/domain/use-cases/save-file.use-case.test.ts
@@ -5,10 +5,22 @@ import { SaveFile } from './save-file.use-case';
 
 describe('SaveFileUseCase', () => {  
 
+  const customOptions = {
+    fileContent: 'custom content',
+    fileDestination: 'custom-outputs/file-destination',
+    fileName: 'custom-table-name',
+  }
+
+  const customFilePath = `${customOptions.fileDestination}/${customOptions.fileName}.txt`;
+
 
   afterEach(() => {
-    fs.rmSync('outputs', {recursive: true});
-    
+    const outputFolderExsist = fs.existsSync('outputs');
+    if (outputFolderExsist) fs.rmSync('outputs', {recursive: true});
+
+    const customOutputFolderExists = fs.existsSync(customOptions.fileDestination);
+    if ( customOutputFolderExists ) fs.rmSync(customOptions.fileDestination, { recursive: true });
+
   })
 
   test('Should save file with default values', () => {
@@ -26,6 +38,7 @@ describe('SaveFileUseCase', () => {
     expect(result).toBe(true);
     expect(fileExsist).toBe(true);
     expect(fileContent).toBe(options.fileContent)
+
 
   });
